+++ src/withPoints/withPoints.c
@@ -200,8 +200,8 @@ _pgr_withpoints(PG_FUNCTION_ARGS) {
 
         int64_t seq = funcctx->call_cntr == 0?  1 : result_tuples[funcctx->call_cntr - 1].start_id;
 
-        values[0] = Int32GetDatum(funcctx->call_cntr + 1);
-        values[1] = Int32GetDatum(seq);
+        values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
+        values[1] = Int32GetDatum((int32_t)seq);
         values[2] = Int64GetDatum(result_tuples[funcctx->call_cntr].start_id);
         values[3] = Int64GetDatum(result_tuples[funcctx->call_cntr].end_id);
         values[4] = Int64GetDatum(result_tuples[funcctx->call_cntr].node);
